Split methods and modules apart
