+++ lib/chef/provisioning/oneview/v1.2/api.rb
@@ -1,0 +1,3 @@
+module OneViewAPIv1_2
+  # TODO
+end
