--- lib/chef/provisioning/oneview/v1.20/api.rb
@@ -1,3 +1,0 @@
-module OneViewAPIv1_20
-  # TODO
-end
