+++ lib/chef/provisioning/rest.rb
