+++ lib/chef/provisioning/version.rb
