+++ spec/support/fixtures/icsp/v102/os-deployment-servers_1670001.json
@@ -1,0 +1,83 @@
+{
+  "type": "OSDServer",
+  "modified": "2015-08-10T04:55:27.017Z",
+  "deviceGroups": null,
+  "customAttributes": [],
+  "created": "2015-08-10T04:35:40.000Z",
+  "uuid": "7d805e1c-beac-e84e-868b-2fb0b6504474",
+  "cpus": [
+    {
+      "cacheSize": "15360",
+      "speed": "1200",
+      "stepping": "7",
+      "model": "Intel(R) Xeon(R) CPU E5-2620 0 @ 2.00GHz",
+      "family": "X64",
+      "status": "ON-LINE",
+      "slot": "0"
+    }
+  ],
+  "ilo": null,
+  "serverLocation": {
+    "rack": "DemoRack",
+    "enclosure": "dazzle",
+    "bay": "2"
+  },
+  "operatingSystemVersion": null,
+  "opswLifecycle": "MANAGED",
+  "osSPVersion": null,
+  "peerIP": "192.168.133.36",
+  "stage": "UNKNOWN",
+  "running": "false",
+  "reporting": true,
+  "storageDevices": [],
+  "ram": "16291780",
+  "swap": "8216568",
+  "architecture": null,
+  "defaultGateway": "172.125.32.1",
+  "discoveredDate": "2015-08-10T04:35:39.000Z",
+  "facility": "Appliance",
+  "hardwareModel": "PROLIANT BL460C GEN8",
+  "jobsHistory": null,
+  "lastScannedDate": null,
+  "loopbackIP": null,
+  "managementIP": "192.168.133.36",
+  "manufacturer": "HP",
+  "mid": "1670001",
+  "netBios": null,
+  "operatingSystem": null,
+  "osFlavor": null,
+  "serialNumber": "VCGE9KB041",
+  "locale": null,
+  "description": null,
+  "status": "OK",
+  "uri": "/rest/os-deployment-servers/1670001",
+  "name": "chef-web01.domain.com",
+  "interfaces": [
+    {
+      "speed": "10000",
+      "netmask": "255.255.254.0",
+      "duplex": "FULL",
+      "ipv6Addr": "",
+      "macAddr": "11:11:11:11:01:15",
+      "dhcpEnabled": false,
+      "ipv4Addr": "172.125.32.38",
+      "type": "ETHERNET",
+      "slot": "eth1"
+    },
+    {
+      "speed": "10000",
+      "netmask": "255.255.255.0",
+      "duplex": "FULL",
+      "ipv6Addr": "",
+      "macAddr": "11:11:11:11:01:14",
+      "dhcpEnabled": true,
+      "ipv4Addr": "192.168.133.36",
+      "type": "ETHERNET",
+      "slot": "eth0"
+    }
+  ],
+  "state": "OK",
+  "hostName": "chef-web01.domain.com",
+  "eTag": "2015-08-10T04:55:27.017Z",
+  "category": "os-deployment-servers"
+}
