+++ spec/support/fixtures/icsp/v102/os-deployment-servers_fakesn.json
@@ -1,0 +1,9 @@
+{
+  "details"=>"",
+  "errorSource"=>nil,
+  "nestedErrors"=>[],
+  "recommendedActions"=>["Verify that the server has not been deleted."],
+  "errorCode"=>nil,
+  "data"=>{},
+  "message"=>"The server ID \"FAKESN\" is not valid."
+}
