+++ spec/unit/icsp_search_spec.rb
@@ -7,7 +7,7 @@ RSpec.describe Chef::Provisioning::OneViewDriver do
 
     it 'gets info for a valid SN' do
       ret_val = @instance.instance_eval { get_icsp_server_by_sn('VCGE9KB041') }
-      expect(ret_val['attributes']['osdServerSerialNumber']).to eq('VCGE9KB041')
+      expect(ret_val['serialNumber']).to eq('VCGE9KB041')
       expect(ret_val['uri']).to match(%r{\/rest\/os-deployment-servers\/.+})
     end
